refactor(CSG): more restructuring: - FuzzyCSGFactory to CSG has been moved to CSG module - modified imports and a few structures
